// src/modules/book/book.interface.ts
export enum Genre {
  FICTION = "FICTION",
  NON_FICTION = "NON_FICTION",
  SCIENCE = "SCIENCE",
  HISTORY = "HISTORY",
  BIOGRAPHY = "BIOGRAPHY",
  FANTASY = "FANTASY"
}

export interface IBook {
    title: string;
    author: string;
    genre: Genre;
    isbn: string;
    description?: string;
    copies: number;
    available: boolean;
    createdAt?: Date;
    updatedAt?: Date;
}

export interface IBookMethods{
   updateAvailability(): void;
}